requirements.txt
django
django-material
django-recaptcha